Separated controllers and routes for survey
